fix(cart drawer): Fix minor issues + improve design
